requirements.txt
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.4
